requirements.txt
Flask
PyJWT
validate_email
